Initial commit
I am including here the template used in the book called OpenGL Temp. The template includes the Debug Properties set up (for my paths) and a main.cpp file that hold a basic skeleton for shaders to be built on.

I am also including Practice 1 from chapter 2.

AND I am including the glew32.dll file that is needed to be store in the built project folder.